+++ cmd/wtm/main.go
@@ -49,7 +49,7 @@ func newRootCmd() *cobra.Command {
 		},
 	}
 	root.AddCommand(
-		newCreateCmd(a), newListCmd(a), newStartCmd(a), newStopCmd(a), newRemoveCmd(a),
+		newCreateCmd(a), newAdoptCmd(a), newListCmd(a), newStartCmd(a), newStopCmd(a), newRemoveCmd(a),
 		newExecCmd(a), newRunCmd(a), newPathCmd(a),
 		newProjectCmd(a), newBackupCmd(a), newDoctorCmd(a),
 	)